Add: ✨ Optimistic Updates
This is to imply updating the state before performing mutation under the
assumption that nothing can go wrong.

It is typically done to give impression that your app is blazing fast.

Just be aware that when dealing optimistic updates, you have to
cater to scenarios where the mutation can in fact error out.

Handling the the optimistic updates is typically not so straight forward.
React query on the other hand does simplify it to a good extent